Urls for submitted_by and assigned_to
Added urls and associated tests to return tickets that were submitted
by a user or assigned to a user.  the My Tickets view now returns
tickets that are both assigned to and submitted by a user.  Tests were
added to ensure that these views return the correct subset of
tickets.  Also added tests to ensure that an appropriate message is
returned if no tickets are included in the query set.  All tests
currently pass.