Updated my_ticket urls to use username instead of id
The original urls used id to identify unique users.  This commit
changes these urls, views and templates to use (or expect) a
username.  This avoids conflicts in the template when we had a user
object who was logged in, and a user object associated with the
selected tickets.  It also makes the urls much more human readable.
All tests continue to pass.