Added title to ticket model
This commit adds a title field to the ticket models. The title is
displayed on the ticket list and ticket detail pages.  Users were
having touble creating simple titles when it was taken from the first
line of the description. This makes it explicit, and hopefully more
intuiative. Several tests where updated to accomodate the new
field. All existing tests continue to pass.