Preparing to upgrade and fixed migration issue.
There are a number of packages that need to be upgraded. This commit
is a restore point in preparation for the upgrade.  There was also an
issue with the local migration that was causing an issue with the
migration history.  Modifying the local, initial migration of tickets
seems to have addressed the problem.