Updated the Readme file and added get_env_variable to settings
This commit removes secret key from the settings file (which was alway
just a placeholder anyway), and replaces it with get_env_variable() as
recommend in 2-scoops of Django.  The readme file was updated to
explain this in more detail and include instructions on setting
secret key. All existing tests continue to pass.