Fixed MyTickets Menu option
The my tickets menu option was providing an intenger to the url rather
than a username.  Once updated, the filters worked as expected and the
ticket list and associated message rendered correctly.  There was a
test, but it was passing when it should have failed. It has been updated
to explicitly include the text of the expected message.  All existing
tests contintue to pass.