Added Sphinx documentation
This commit includes the source file for basic sphinx documentation that
will make the existing doc strings in the models, views, forms, and
utilities available in a single place.  The readme has been updated to
include some brief instructions on rebuilding the documentation.  All
previously passing tests continue to pass.