Disable vote button if user has already voted
This commit includes logic to disable the vote button on the ticket
detail page if the user has already voted for this ticket.  It also
includes a warning if that you can only vote once if you try to access
the url directly. The vote-twice unit test was updated to ensure the
warning renders appropriately.