Fixed default redirect url and made search required
This is a small commit fix a couple of identified issues.  The text box
on the navbar is now required to prevent searching for nothing and
redirecting to the ticket list.  Additionally, the redirect url after
login has been updated to return to the ticket list.  All existing tests
continue to pass.