Final updates to templates associated migration to bootstrap5
This commit includes the final updates associated with the migration
to bootstrap5. Specifically it updates the tests (or files that were
idetifed by failing tests). It also includes a small update to the
jquery associated with tag list to update the tag count to reflect
currently visible tickets.  All previoulsy passing tests continue to pass.