Fixes for missing main-content anchor, Skip to Filter Menu added, Styling changes
Main content anchor added on pages that were missing it, such as pages containing forms.

Skip to filter button added along on templates that contain the ticket filter.

Asterisk styling added to base template so that they are not repeated on templates that use it (form templates since asterisks are used to indicate required fields).